+++ vss_cli/config.py
@@ -1227,6 +1227,7 @@ class Configuration(VssManager):
         try:
             from vss_cli.plugins.compute_plugins.callbacks import (
                 process_user_data,
+                process_day_zero,
             )
 
             spec_payload = dict()
@@ -1350,6 +1351,23 @@ class Configuration(VssManager):
                     ] = validate_json_file_or_type(
                         self, 'additional_parameters', add_params
                     )
+                # day_zero
+                day_zero = payload.get('day-zero')
+                if day_zero is not None:
+                    _d0 = process_day_zero(
+                        self, 'day-zero', day_zero['config']
+                    )
+                    day_zero_payload = {
+                        'config': _d0[0],
+                        'config-encoding': _d0[1],
+                    }
+                    if 'id-token' in day_zero:
+                        _i0 = process_day_zero(
+                            self, 'id-token', day_zero['id-token']
+                        )
+                        _i0_d = {'idtoken': _i0[0], 'idtoken-encoding': _i0[1]}
+                        day_zero_payload.update(_i0_d)
+                    spec_payload['day_zero'] = day_zero_payload
             if built == 'os_install':
                 spec_payload['iso'] = self.get_iso_by_name_or_path(
                     machine_section['iso']